feat: standardize reducer utilities and consolidate in ListWithFilters
- Move pendingFilters.ts content to ListWithFilters.reducerUtils.ts
- Rename PendingFiltersState to ListWithFiltersState
- Consolidate common actions as ListWithFiltersActions enum
- Make groupFn optional (returns empty Map when not provided)
- Update all 7 reducers to import from consolidated location
- Remove Collections no-op groupFn since grouping isn't used

This improves code organization by keeping ListWithFilters-related
utilities together and reduces duplication across reducers.
