+++ src/components/Reviews/reducer.ts
@@ -8,19 +8,18 @@ import {
   buildGroupValues,
   clearPendingFilters,
   createInitialState,
-  type PendingFiltersState,
-  resetPendingFilters,
-  showMore,
-  updatePendingFilter,
-  updateSort,
-} from "~/utils/pendingFilters";
-import {
   createReleaseYearFilter,
   createTitleFilter,
   getGroupLetter,
+  ListWithFiltersActions,
+  type ListWithFiltersState,
+  resetPendingFilters,
+  showMore,
   sortNumber,
   sortString,
-} from "~/utils/reducerUtils";
+  updatePendingFilter,
+  updateSort,
+} from "~/components/ListWithFilters.reducerUtils";
 
 const SHOW_COUNT_DEFAULT = 100;
 
@@ -35,16 +34,16 @@ type ReviewsSort =
   | "title-desc";
 
 export const Actions = {
-  APPLY_PENDING_FILTERS: "APPLY_PENDING_FILTERS",
-  CLEAR_PENDING_FILTERS: "CLEAR_PENDING_FILTERS",
+  APPLY_PENDING_FILTERS: ListWithFiltersActions.APPLY_PENDING_FILTERS,
+  CLEAR_PENDING_FILTERS: ListWithFiltersActions.CLEAR_PENDING_FILTERS,
   PENDING_FILTER_GENRES: "PENDING_FILTER_GENRES",
   PENDING_FILTER_GRADE: "PENDING_FILTER_GRADE",
   PENDING_FILTER_RELEASE_YEAR: "PENDING_FILTER_RELEASE_YEAR",
   PENDING_FILTER_REVIEW_YEAR: "PENDING_FILTER_REVIEW_YEAR",
   PENDING_FILTER_TITLE: "PENDING_FILTER_TITLE",
-  RESET_PENDING_FILTERS: "RESET_PENDING_FILTERS",
-  SHOW_MORE: "SHOW_MORE",
-  SORT: "SORT",
+  RESET_PENDING_FILTERS: ListWithFiltersActions.RESET_PENDING_FILTERS,
+  SHOW_MORE: ListWithFiltersActions.SHOW_MORE,
+  SORT: ListWithFiltersActions.SORT,
 } as const;
 
 export type ActionType =
@@ -108,7 +107,7 @@ type SortAction = {
   value: ReviewsSort;
 };
 
-type State = PendingFiltersState<ReviewsListItemValue, ReviewsSort>;
+type State = ListWithFiltersState<ReviewsListItemValue, ReviewsSort>;
 
 // Helper functions
 function getReviewDateGroup(value: ReviewsListItemValue): string {
