--- src/utils/reducerUtils.ts
@@ -1,78 +1,0 @@
-/**
- * Common reducer utilities and filter handlers.
- *
- * These functions provide reusable logic for common reducer patterns
- * like filtering, sorting, and pagination while maintaining full type safety.
- *
- * Uses consistent generic type parameters:
- * - TItem: The type of items in the list
- * - TSortValue: The type of sort values
- * - TGroupedValues: The type of grouped values structure
- */
-import { collator } from "./collator";
-
-export function createNameFilter(value: string) {
-  const regex = new RegExp(value, "i");
-  return <T extends { name: string }>(item: T) => regex.test(item.name);
-}
-
-export function createReleaseYearFilter(minYear: string, maxYear: string) {
-  return <T extends { releaseYear: string }>(item: T) => {
-    return item.releaseYear >= minYear && item.releaseYear <= maxYear;
-  };
-}
-
-export function createReviewYearFilter(minYear: string, maxYear: string) {
-  return <T extends { reviewYear?: string }>(item: T) => {
-    const year = item.reviewYear;
-    if (!year) return false;
-    return year >= minYear && year <= maxYear;
-  };
-}
-
-// Common filter creators - simple functions that create filter functions
-export function createTitleFilter(value: string) {
-  const regex = new RegExp(value, "i");
-  return <T extends { title: string }>(item: T) => regex.test(item.title);
-}
-
-// Filter values helper
-export function filterValues<TItem>({
-  filters,
-  values,
-}: {
-  filters: Record<string, (arg0: TItem) => boolean>;
-  values: readonly TItem[];
-}): TItem[] {
-  return values.filter((item) => {
-    return Object.values(filters).every((filter) => {
-      return filter(item);
-    });
-  });
-}
-
-/**
- * Gets the group letter for a given string, typically used for alphabetical grouping.
- * Non-alphabetic characters are grouped under "#".
- *
- * @param str - The string to get the group letter from
- * @returns The uppercase first letter or "#" for non-alphabetic characters
- */
-export function getGroupLetter(str: string): string {
-  const letter = str.slice(0, 1);
-
-  // Check if the character is non-alphabetic (same in upper and lower case)
-  if (letter.toLowerCase() === letter.toUpperCase()) {
-    return "#";
-  }
-
-  return letter.toLocaleUpperCase();
-}
-
-export function sortNumber(a: number, b: number): number {
-  return a - b;
-}
-
-export function sortString(a: string, b: string): number {
-  return collator.compare(a, b);
-}
